requirements.txt
black==20.8b1
grpcio==1.27.2
flake8==3.8.4
flask==1.1.1
Flask-Cors==3.0.8
fuzzysearch==0.7.0
fuzzywuzzy==0.18.0
jsonpickle==1.4.2
nltk==3.4.5
num2words==0.5.10
pony==0.7.14
protobuf==3.11.3
python-Levenshtein==0.12.0
pyyaml==5.3.1
redis==3.5.3
requests==2.23.0
scikit-learn==0.22.2.post1
wrapt_timeout_decorator==1.3.1
